+++ src/ralph/backends/storage/__init__.py
@@ -1,0 +1,1 @@
+"""Storage backends for Ralph"""